+++ cms/src/components/header/Header.tsx
@@ -1,21 +1,18 @@
-import React, { useEffect } from "react";
+import React from "react";
 
 import NavAnchor from "../navAnchor/NavAnchor";
 import styles from "./Header.module.css";
 import { useAuth } from "../../hooks/useAuth";
-import { usePrevLocation } from "../../hooks/usePrevLocation";
-import { useUserData } from "../../hooks/useUser";
 import { Link, useLocation, useNavigate } from "@tanstack/react-router";
 import router from "../../config/router.config";
 
 const Header: React.FC = () => {
   const { accessToken, isAuthenticated, logout } = useAuth();
+  const navigate = useNavigate();
   const location = useLocation();
   const from = location.pathname;
-  const navigate = useNavigate();
   console.group("Header component rendering...");
-  // console.log("prevLocation in Header component:", prevLocation);
-  // console.log("from in Header component:", from);
+  console.log("from in Header component:", from);
   console.log("accessToken in Header component:", accessToken);
   console.log("isAuthenticated in Header component:", isAuthenticated);
   console.groupEnd();
@@ -74,7 +71,7 @@ const Header: React.FC = () => {
               </li>
 
               <li>
-                <NavAnchor pathname="/signup" textContent="sign up" />
+                <NavAnchor pathname="/signup" textContent="create account" />
               </li>
             </>
           )}
@@ -82,37 +79,6 @@ const Header: React.FC = () => {
       </nav>
     </header>
   );
-  /* return (
-    <header>
-      <nav>
-        <ul className={styles["nav-left"]}>
-          <li>
-            <Link to="/">
-              <h1>Project Name</h1>
-            </Link>
-          </li>
-        </ul>
-
-        <ul className={styles["nav-right"]}>
-          <li>
-            <NavAnchor pathname="/" textContent="home" />
-          </li>
-
-          <li>
-            <NavAnchor pathname="/faq" textContent="faq" />
-          </li>
-
-          <li>
-            <NavAnchor pathname="/login" textContent="login" />
-          </li>
-
-          <li>
-            <NavAnchor pathname="/signup" textContent="sign up" />
-          </li>
-        </ul>
-      </nav>
-    </header>
-  ); */
 };
 
 export default Header;