+++ cms/src/features/login/Login.tsx
@@ -1,23 +1,22 @@
 import NavAnchor from "../../components/navAnchor/NavAnchor";
 import LoginForm from "./components/LoginForm";
 import { useAuth } from "../../hooks/useAuth";
-import { usePrevLocation } from "../../hooks/usePrevLocation";
-import { Navigate, useLocation, useSearch } from "@tanstack/react-router";
-import router from "../../config/router.config";
+import { Navigate, useSearch } from "@tanstack/react-router";
 
 const Login = () => {
-  const { accessToken } = useAuth();
+  const { accessToken, isAuthenticated } = useAuth();
   // const location = useLocation();
   const searchParams = useSearch({
     from: "/login",
     select: (params: { redirect?: string }) => params,
   });
   console.group("Login component rendering...");
+  console.log("accessToken:", accessToken);
+  console.log("isAuthenticated:", isAuthenticated);
   // console.log("prevLocation:", prevLocation);
   // If accessToken exists (user is logged in)
   //  Navigate the user to their dashboard or the previous location
   if (accessToken) {
-    console.log("[Login component] location:", location);
     console.log("[Login component] searchParams:", searchParams);
     console.log(
       "[Login component] searchParams.redirect:",
@@ -25,20 +24,18 @@ const Login = () => {
     );
     const redirectTo = searchParams.redirect ?? "/dashboard";
     console.log("redirectTo:", redirectTo);
-    // return <Navigate to={"/dashboard"} href={redirectTo} replace={true} />;
+    return <Navigate to={"/dashboard"} href={redirectTo} replace={true} />;
   }
-
-  const redirectTo = searchParams.redirect ?? "/dashboard";
 
   return (
     <section>
-      <h2>Sign in</h2>
+      <h2>Login</h2>
 
-      <LoginForm redirectTo={redirectTo} />
+      <LoginForm />
 
       <p>
         <span>New here? </span>
-        <NavAnchor pathname="/signup" textContent="Sign up" />
+        <NavAnchor pathname="/signup" textContent="Create account" />
       </p>
 
       <p>