+++ cms/src/routes/_protected/_dashboard/_user/$username/posts.tsx
@@ -1,9 +1,17 @@
-import { createFileRoute } from "@tanstack/react-router";
+import { createFileRoute, useRouteContext } from "@tanstack/react-router";
 
-export const Route = createFileRoute("/_protected/_dashboard/_user/$username/posts")({
+export const Route = createFileRoute(
+  "/_protected/_dashboard/_user/$username/posts"
+)({
   component: RouteComponent,
 });
 
 function RouteComponent() {
+  const context = useRouteContext({
+    from: "/_protected",
+  });
+  console.group("[Posts] component rendering...");
+  console.log("context:", context);
+  console.groupEnd();
   return <div>Hello "/_protected/$username/posts"!</div>;
 }